third_party/blink/web_tests/fast/lists/ol-reversed-nested-items-expected.txt

This tests that list items are properly marked when some of the items are indirect children of a reversed list.

12 Twelve
11 Eleven
10 Ten
9 Nine
8 Eight
-> 7 Seven
4 Four
3 Three
-> 2 Two
-> 1 One
0 Zero
-1 Minus One
